Fix files for single class
