+++ src/postgres-mcp-server/awslabs/postgres_mcp_server/connection/rds_connector.py
@@ -19,7 +19,7 @@ import boto3
 from loguru import logger
 from typing import Any, Dict, List, Optional
 
-from awslabs.postgres_mcp_server.connection.connection_factory import AbstractDBConnection
+from awslabs.postgres_mcp_server.connection.abstract_class import AbstractDBConnection
 
 
 class RDSDataAPIConnection(AbstractDBConnection):
