+++ src/postgres-mcp-server/awslabs/postgres_mcp_server/server.py
@@ -354,7 +354,7 @@ def main():
 
     logger.success('Successfully validated database connection to Postgres')
 
-    logger.info('Started Postgres MCP server')
+    logger.info('Starting Postgres MCP server')
     mcp.run()
